Remove redundant cast

+++ zmij.cc
@@ -715,9 +715,9 @@ ZMIJ_INLINE auto to_unshuffled_digits(uint64_t value, const constants& c)
 
   // When reverse_hi_lo, actual order is abbccddee|ffgghhii; the 17th digit
   // ends up at byte index 0, so it can be extracted for the trailing write.
-  uint64x1_t ffgghhii_bbccddee_64 = {
-      reverse_hi_lo ? (uint64_t(abbccddee) << 32) | ffgghhii
-                    : (uint64_t(ffgghhii) << 32) | abbccddee};
+  uint64x1_t ffgghhii_bbccddee_64 = {reverse_hi_lo
+                                         ? (abbccddee << 32) | ffgghhii
+                                         : (ffgghhii << 32) | abbccddee};
   int32x2_t bbccddee_ffgghhii = vreinterpret_s32_u64(ffgghhii_bbccddee_64);
 
   int32x2_t bbcc_ffgg = vreinterpret_s32_u32(
@@ -729,7 +729,6 @@ ZMIJ_INLINE auto to_unshuffled_digits(uint64_t value, const constants& c)
 
   int32x4_t ddee_bbcc_hhii_ffgg = vreinterpretq_s32_u32(
       vshll_n_u16(vreinterpret_u16_s32(ddee_bbcc_hhii_ffgg_32), 0));
-
   return to_bcd_4x4(ddee_bbcc_hhii_ffgg, c);
 }
